+++ src/retrieve_players/mapping.py
@@ -1,9 +1,9 @@
 from commands.types import CommandMapping
-from retrieve_players.retrieve_action import retrieve_player_list
+import retrieve_players.retrieve_players_main as retrieve_players_main
 
 retrieve_commands: CommandMapping = {
     "retrieve_players": {
-        "function": retrieve_player_list,
+        "function": retrieve_players_main.retrieve_player_list,
         "description": "Retrieves list of players of an event",
         "params": ["startgg_link"]
     }
